+++ inst/NEWS.md
@@ -1,6 +1,7 @@
 # fastgssa 0.2.1 (04-07-2019)
 
 * BUG FIX: Fix isinf scope issue for Windows users.
+* MINOR CHANGE: Move ggplot2 to Suggests
 
 # fastgssa 0.2.0 (21-06-2019)